+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.1]
+
+### Added
+
+- Biconvex Pie Chart shape implementation
+
+### Changed
+
+- Deprecated AreaPlot, LinePlot, PolarPlotSeries, StairstepPlot to replace with AreaPlot2, LinePlot2, PolarPlotSeries2,
+  StairstepPlot2 as part of deprecation of HoverableElementArea.
+
 ## [0.10.0]
 
 ### Added